build: rename `index.joke` to `guestbook.joke`
This is a more meaningful name.

A user will likely have to edit and rename the file anyway.
(For example, I rename it to `index.cgi` on SDF.)
We can't force them to use a "cool" directory URI
(https://www.w3.org/Provider/Style/URI)
with a default filename.
The new name lets them have their own `index.joke`
in the same directory.